Introduce package struct to align API calls output
To make sure the different API calls which expose package information (/list, /package/*) have the same content / naming, the package object is introduced.

The icon path is added to the detail output to have in the detail view all the information which is also on the top level.

Moved handler and to it's own file for better readability. Server now must be started with `go run .` instead of `go run main.go`.