Update main.go to have nicer shutdown mechanism
This refactoring initially came from the issue that `Ctrl-C` did not stop the docker container. It turns out to make it work the container needs to be run with `-i -t`, but the refactoring I did I think is still helpful.

To make params work with Docker like `--address`, ENTRYPOINT instead of CMD is now used for the mandatory parts.